# openapi/artifacts/assistant-api.yaml
openapi: 3.0.3
info:
  title: Rapida Assistant API
  version: 0.1.0
paths:
...
  /v1/assistant-deployment/get-phone-deployment/{assistantId}:
    get:
      operationId: getAssistantPhoneDeployment
      summary: Get assistant phone deployment
      description: Gets the latest phone deployment for an assistant. This endpoint requires user authentication with organization and project scope.
      parameters:
        - name: assistantId
          in: path
          required: true
          schema:
            $ref: "./common.yaml#/components/schemas/Uint64String"
      responses:
        "200":
          description: Assistant phone deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAssistantPhoneDeploymentResponse"
        "400":
          description: Invalid assistantId
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated request
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
        "403":
          description: Missing authentication scope
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
        "500":
          description: Failed to get assistant phone deployment
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
  /v1/assistant-deployment/get-webplugin-deployment/{assistantId}:
    get:
      operationId: getAssistantWebpluginDeployment
      summary: Get assistant webplugin deployment
      description: Gets the latest webplugin deployment for an assistant. This endpoint requires user authentication with organization and project scope.
      parameters:
        - name: assistantId
          in: path
          required: true
          schema:
            $ref: "./common.yaml#/components/schemas/Uint64String"
      responses:
        "200":
          description: Assistant webplugin deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAssistantWebpluginDeploymentResponse"
        "400":
          description: Invalid assistantId
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated request
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
        "403":
          description: Missing authentication scope
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
        "500":
          description: Failed to get assistant webplugin deployment
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    GetAssistantPhoneDeploymentResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/AssistantPhoneDeployment"
        error:
          $ref: "./common.yaml#/components/schemas/Error"
...
    GetAssistantWebpluginDeploymentResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/AssistantWebpluginDeployment"
        error:
          $ref: "./common.yaml#/components/schemas/Error"
...
    AssistantPhoneDeployment:
      type: object
      properties:
        id:
          $ref: "./common.yaml#/components/schemas/Uint64String"
        assistantId:
          $ref: "./common.yaml#/components/schemas/Uint64String"
        greeting:
          type: string
        greetingInterruptible:
          type: boolean
        mistake:
          type: string
        inputAudio:
          $ref: "#/components/schemas/DeploymentAudioProvider"
        outputAudio:
          $ref: "#/components/schemas/DeploymentAudioProvider"
        phoneProviderName:
          type: string
        phoneOptions:
          type: array
          items:
            $ref: "./common.yaml#/components/schemas/Metadata"
        status:
          type: string
        maxSessionDuration:
          type: integer
          format: uint64
          minimum: 180
          maximum: 600
        idealTimeout:
          type: integer
          format: uint64
          minimum: 15
          maximum: 120
        idealTimeoutBackoff:
          type: integer
          format: uint64
          minimum: 0
          maximum: 5
        idealTimeoutMessage:
          type: string
...
    AssistantWebpluginDeployment:
      type: object
      properties:
        id:
          $ref: "./common.yaml#/components/schemas/Uint64String"
        assistantId:
          $ref: "./common.yaml#/components/schemas/Uint64String"
        greeting:
          type: string
        greetingInterruptible:
          type: boolean
        mistake:
          type: string
        inputAudio:
          $ref: "#/components/schemas/DeploymentAudioProvider"
        outputAudio:
          $ref: "#/components/schemas/DeploymentAudioProvider"
        suggestion:
          type: array
          items:
            type: string
        status:
          type: string
        maxSessionDuration:
          type: integer
          format: uint64
          minimum: 180
          maximum: 600
        idealTimeout:
          type: integer
          format: uint64
          minimum: 15
          maximum: 120
        idealTimeoutBackoff:
          type: integer
          format: uint64
          minimum: 0
          maximum: 5
        idealTimeoutMessage:
          type: string
...
    DeploymentAudioProvider:
      type: object
      properties:
        id:
          $ref: "./common.yaml#/components/schemas/Uint64String"
        audioType:
          type: string
        audioProvider:
          type: string
        audioOptions:
          type: array
          items:
            $ref: "./common.yaml#/components/schemas/Metadata"
        status:
          type: string
...